% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misszx.R
\name{misszx}
\alias{misszx}
\title{Outpatient log statistics}
\usage{
misszx(data)
}
\arguments{
\item{data}{Datasets that need to be analyzed}
}
\value{
Return the statistical data table
}
\description{
Statistics on the missing logs of outpatient clinics
}
\examples{
misszx(data)
}
